Initial structure of the Seminarjs demo
This still requires Seminarjs to exist as an npm package.
At the moment, if you want to test it, clone Seminarjs, browse to its
root directory and use `npm link`, then go to this repository's root
directory, and enter `npm link seminarjs`.
That will add seminarjs as a dependency, linking it from its original
folder, so changes take effect live.